create a unified file for game api classes This is instead of having them only in PRC API. Makes more sense and is apart of the Codebase Cleanup Commitment
